+++ truffle/CHANGELOG.md
@@ -52,6 +52,8 @@ This changelog summarizes major changes between Truffle versions relevant to lan
 	* Added `removeHashEntry(Object, Object)` removing the mapping for a given key.
 	* Added `isHashEntryExisting(Object, Object)` that allows to specify that that mapping for a given key is existing.
 	* Added `getHashEntriesIterator(Object)` to return the hash entries iterator.
+    * Added `getHashKeysIterator(Object)` to return the hash keys iterator.
+    * Added `getHashValuesIterator(Object)` to return the hash values iterator.
 
 ## Version 21.0.0
 * If an `AbstractTruffleException` is thrown from the `ContextLocalFactory`, `ContextThreadLocalFactory` or event listener, which is called during the context enter, the exception interop messages are executed without a context being entered. The event listeners called during the context enter are: